refactor: logout
